inference-setup/requirements.txt
pillow
tqdm
torch
pandas
peft
torchvision
